+++ src/transformers/models/__init__.py
@@ -65,7 +65,6 @@ from . import (
     deprecated,
     deta,
     detr,
-    rt_detr,
     dialogpt,
     dinat,
     dinov2,
