+++ tests/models/rt_detr/test_image_processing_rt_detr.py
@@ -12,13 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import unittest
+
 import requests
 
-from transformers.testing_utils import (
-    require_torch,
-    require_vision,
-    slow
-)
+from transformers.testing_utils import require_torch, require_vision, slow
 from transformers.utils import is_torch_available, is_vision_available
 
 from ...test_image_processing_common import ImageProcessingTestMixin, prepare_image_inputs
@@ -78,10 +75,12 @@ class RtDetrImageProcessingTester(unittest.TestCase):
             torchify=torchify,
         )
 
+
 @require_torch
 @require_vision
 class RtDetrImageProcessingTest(ImageProcessingTestMixin, unittest.TestCase):
     image_processing_class = RtDetrImageProcessor if is_vision_available() else None
+
     def setUp(self):
         self.image_processor_tester = RtDetrImageProcessingTester()
 
@@ -91,16 +90,15 @@ class RtDetrImageProcessingTest(ImageProcessingTestMixin, unittest.TestCase):
 
     def test_image_processor_properties(self):
         image_processing = self.image_processing_class(**self.image_processor_dict)
-        self.assertTrue(hasattr(image_processing, "format"))
         self.assertTrue(hasattr(image_processing, "do_resize"))
         self.assertTrue(hasattr(image_processing, "size"))
         self.assertTrue(hasattr(image_processing, "resample"))
         self.assertTrue(hasattr(image_processing, "do_rescale"))
         self.assertTrue(hasattr(image_processing, "rescale_factor"))
         self.assertTrue(hasattr(image_processing, "return_tensors"))
-        
+
     @slow
-    def test_image_processor_outputs(self):        
+    def test_image_processor_outputs(self):
         image = Image.open("./tests/fixtures/tests_samples/COCO/000000039769.png")
 
         image_processing = self.image_processing_class(**self.image_processor_dict)
@@ -115,25 +113,27 @@ class RtDetrImageProcessingTest(ImageProcessingTestMixin, unittest.TestCase):
         self.assertTrue(torch.allclose(encoding["pixel_values"][0, 0, 0, :3], expected_slice, atol=1e-5))
 
     @slow
-    def test_multiple_images_processor_outputs(self):       
-        images_urls = ["http://images.cocodataset.org/val2017/000000000139.jpg",
+    def test_multiple_images_processor_outputs(self):
+        images_urls = [
+            "http://images.cocodataset.org/val2017/000000000139.jpg",
             "http://images.cocodataset.org/val2017/000000000285.jpg",
             "http://images.cocodataset.org/val2017/000000000632.jpg",
             "http://images.cocodataset.org/val2017/000000000724.jpg",
             "http://images.cocodataset.org/val2017/000000000776.jpg",
             "http://images.cocodataset.org/val2017/000000000785.jpg",
             "http://images.cocodataset.org/val2017/000000000802.jpg",
-            "http://images.cocodataset.org/val2017/000000000872.jpg",]
-        
+            "http://images.cocodataset.org/val2017/000000000872.jpg",
+        ]
+
         images = []
         for url in images_urls:
             image = Image.open(requests.get(url, stream=True).raw)
             images.append(image)
-        
-        # apply image processing        
+
+        # apply image processing
         image_processing = self.image_processing_class(**self.image_processor_dict)
         encoding = image_processing(images=images, return_tensors="pt")
-        
+
         # verify if pixel_values is part of the encoding
         self.assertIn("pixel_values", encoding)
 
@@ -142,13 +142,16 @@ class RtDetrImageProcessingTest(ImageProcessingTestMixin, unittest.TestCase):
         self.assertEqual(encoding["pixel_values"].shape, expected_shape)
 
         # verify pixel values: output values
-        expected_slices = torch.tensor([[0.5333333611488342, 0.5568627715110779, 0.5647059082984924],
-            [0.5372549295425415, 0.4705882668495178, 0.4274510145187378],
-            [0.3960784673690796, 0.35686275362968445, 0.3686274588108063],
-            [0.20784315466880798, 0.1882353127002716, 0.15294118225574493],
-            [0.364705890417099, 0.364705890417099, 0.3686274588108063],
-            [0.8078432083129883, 0.8078432083129883, 0.8078432083129883],
-            [0.4431372880935669, 0.4431372880935669, 0.4431372880935669],
-            [0.19607844948768616, 0.21176472306251526, 0.3607843220233917]])
+        expected_slices = torch.tensor(
+            [
+                [0.5333333611488342, 0.5568627715110779, 0.5647059082984924],
+                [0.5372549295425415, 0.4705882668495178, 0.4274510145187378],
+                [0.3960784673690796, 0.35686275362968445, 0.3686274588108063],
+                [0.20784315466880798, 0.1882353127002716, 0.15294118225574493],
+                [0.364705890417099, 0.364705890417099, 0.3686274588108063],
+                [0.8078432083129883, 0.8078432083129883, 0.8078432083129883],
+                [0.4431372880935669, 0.4431372880935669, 0.4431372880935669],
+                [0.19607844948768616, 0.21176472306251526, 0.3607843220233917],
+            ]
+        )
         self.assertTrue(torch.allclose(encoding["pixel_values"][:, 1, 0, :3], expected_slices, atol=1e-5))
-
